refactor & fixed some ui related isssue

+++ public/event.php
@@ -12,7 +12,6 @@ if ($_SERVER['REQUEST_METHOD'] === "POST") {
 
   // save to db
   query($db, "INSERT INTO events (title, day, month, year) VALUES (:title, :day, :month, :year)", ['title' => $title, 'day' => (int)$eventDay, 'month' => (int)$eventMonth, 'year' => (int)$eventYear]);
-  // header("location:/");
 }
 ?>
 
@@ -29,15 +28,21 @@ if ($_SERVER['REQUEST_METHOD'] === "POST") {
     $lists = query($db, "SELECT id, title from events WHERE day = $eventDay AND month = $eventMonth AND year = $eventYear");
 
     ?>
-    <p style="font-size: 1.4rem; text-transform: capitalize;">list of events</p>
-    <ul style="display: grid; grid-template-columns: repeat(5, 1fr); gap:1rem; border: 2px dotted blue;  padding: 1.5rem;">
-      <?php foreach ($lists as $list): ?>
-        <li style="width:10rem; display: flex; border:1px dotted green;">
-          <p style="font-size: 1.2rem; "><?= $list->title ?></p>
-          <a href="/delete.php?id=<?= $list->id ?>" style="margin-left: auto; align-self: center; padding: .4rem .5rem;">delete</a>
-        </li>
-      <?php endforeach ?>
-    </ul>
+
+    <?php if (count($lists) > 0) : ?>
+      <p style="font-size: 1.4rem; text-transform: capitalize;">list of events</p>
+      <ul style="display: grid; grid-template-columns: repeat(5, 1fr); gap:1rem; border: 2px dotted blue;  padding: 1.5rem;">
+        <?php foreach ($lists as $list): ?>
+          <li style="width:10rem; display: flex; border:1px dotted green;">
+            <p style="font-size: 1.2rem; "><?= $list->title ?></p>
+            <a href="/delete.php?id=<?= $list->id ?>" style="margin-left: auto; align-self: center; padding: .4rem .5rem;">delete</a>
+          </li>
+        <?php endforeach ?>
+      </ul>
+
+    <?php else: ?>
+      <p style="text-align: center; font-size:1.4rem;text-transform: uppercase; padding: 2rem 0; font-family: Arial, Helvetica, sans-serif; font-style: italic; color: gray;">there is no event to display</p>
+    <?php endif ?>
   </div>
 
   <div style="grid-row: 3 / 4; grid-column: 4 / 10; margin-top:1rem;">
